Extract AWS pricing and networking into separate classes
